--- internal/store/file/debug.go
@@ -1,7 +1,0 @@
-package file
-
-import "github.com/davecgh/go-spew/spew"
-
-func (box *Box) Debug() {
-	spew.Dump(box)
-}
