--- internal/store/file/delete.go
@@ -1,5 +1,0 @@
-package file
-
-func (box *Box) Delete(key string) error {
-	return nil
-}
