--- internal/store/memory/debug.go
@@ -1,7 +1,0 @@
-package memory
-
-import "github.com/davecgh/go-spew/spew"
-
-func (box *Box) Debug() {
-	spew.Dump(box)
-}
