--- internal/store/memory/delete.go
@@ -1,5 +1,0 @@
-package memory
-
-func (box *Box) Delete(key string) error {
-	return nil
-}
