--- internal/store/pg/debug.go
@@ -1,7 +1,0 @@
-package pg
-
-import "github.com/davecgh/go-spew/spew"
-
-func (box *Box) Debug() {
-	spew.Dump(box)
-}
