--- internal/store/pg/delete.go
@@ -1,5 +1,0 @@
-package pg
-
-func (box *Box) Delete(key string) error {
-	return nil
-}
