--- internal/store/pg/get.go
@@ -1,16 +1,0 @@
-package pg
-
-import "log"
-
-func (box *Box) Get(key string) string {
-
-	var originalURL string
-	err := box.db.QueryRow("SELECT original_url FROM  short  WHERE  short_url = $1",
-		key,
-	).Scan(&originalURL)
-
-	log.Println("err: ", err)
-	log.Println(originalURL)
-	return originalURL
-
-}
